consolidation of tests
